4.23 done

+++ part4/controllers/blogs.js
@@ -1,25 +1,34 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response, next) => {
-  const body = request.body;
-
-  const user = await User.findById(body.userId);
-  const blog = new Blog({
-    title: body.title,
-    author: body.author || "",
-    url: body.url,
-    likes: body.likes || 0,
-    user: user,
-  });
-
+blogsRouter.post("/", async (request, response, next) => {  
   try {
+    const body = request.body;
+    const user = request.user;
+
+    if (!body.title || !body.url) {
+      return response.status(400).json({ error: "title or url missing" });
+    }
+
+    if(!user) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+  
+    const blog = new Blog({
+      title: body.title,
+      author: body.author || "",
+      url: body.url,
+      likes: body.likes || 0,
+      user: user,
+    });
     const savedBlog = await blog.save();
     user.blogs = user.blogs.concat(savedBlog._id);
     await user.save();
@@ -46,8 +55,15 @@ blogsRouter.put("/:id", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const user = request.user;
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (user.id.toString() === blogToDelete.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: "Unauthorized" });
+  }
 });
 
 module.exports = blogsRouter;
